Move object related types to its own module
