+++ lvgl/src/lv_core/mod.rs
@@ -1,1 +1,5 @@
+#[macro_use]
+mod object;
 pub mod style;
+
+pub use object::*;
